[./] 2014-07-25  Basile Starynkevitch  <basile@starynkevitch.net>
	* INSTALL/README-MELT-PLUGIN: Don't use utfcpp.  Mention new
	features of MELT 1.1..

[gcc/]
2014-07-25  Basile Starynkevitch  <basile@starynkevitch.net>

	* melt-runtime.cc (melt_really_initialize): Improve debug message.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/melt-branch@213048 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/melt-runtime.cc
@@ -10996,7 +10996,8 @@ melt_really_initialize (const char* pluginame, const char*versionstr)
       else
 	curmod.assign(modstr);
       melt_asked_modes_vector.push_back (curmod);
-      debugeprintf("melt_really_initialize curmod='%s' #%d", curmod.c_str(), melt_asked_modes_vector.size());
+      debugeprintf("melt_really_initialize curmod='%s' #%d", curmod.c_str(),
+		   (int)melt_asked_modes_vector.size());
     }
   }
